code cleanup and phpcs fixes
