+++ bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/Business/Reader/ProductListReader.php
@@ -19,9 +19,8 @@ class ProductListReader implements ProductListReaderInterface
     protected $searchQuoteQueryExpanderPlugins;
 
     /**
-     * @pram array<\FondOfOryx\Zed\ProductListSearchRestApiExtension\Dependency\Plugin\SearchQuoteQueryExpanderPluginInterface> $searchQuoteQueryExpanderPlugins
-     *
      * @param \FondOfOryx\Zed\ProductListSearchRestApi\Persistence\ProductListSearchRestApiRepositoryInterface $repository
+     * @pram array<\FondOfOryx\Zed\ProductListSearchRestApiExtension\Dependency\Plugin\SearchQuoteQueryExpanderPluginInterface> $searchQuoteQueryExpanderPlugins
      */
     public function __construct(
         ProductListSearchRestApiRepositoryInterface $repository,
