+++ bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/Communication/Controller/GatewayController.php
@@ -6,7 +6,7 @@ use Generated\Shared\Transfer\ProductListCollectionTransfer;
 use Spryker\Zed\Kernel\Communication\Controller\AbstractGatewayController;
 
 /**
- * @method \FondOfOryx\Zed\ProductListSearchRestApi\Persistence\ProductListSearchRestApiRepository getRepository()
+ * @method \FondOfOryx\Zed\ProductListSearchRestApi\Business\ProductListSearchRestApiFacadeInterface getFacade()
  */
 class GatewayController extends AbstractGatewayController
 {
@@ -18,6 +18,6 @@ class GatewayController extends AbstractGatewayController
     public function searchProductListAction(
         ProductListCollectionTransfer $productListCollectionTransfer
     ): ProductListCollectionTransfer {
-        return $this->getRepository()->searchProductList($productListCollectionTransfer);
+        return $this->getFacade()->findProductLists($productListCollectionTransfer);
     }
 }
