+++ bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/Persistence/ProductListSearchRestApiPersistenceFactory.php
@@ -4,6 +4,10 @@ namespace FondOfOryx\Zed\ProductListSearchRestApi\Persistence;
 
 use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\Mapper\ProductListMapper;
 use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\Mapper\ProductListMapperInterface;
+use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\QueryBuilder\ProductListSearchFilterFieldQueryBuilder;
+use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\QueryBuilder\ProductListSearchFilterFieldQueryBuilderInterface;
+use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\QueryBuilder\QuoteQueryJoinQueryBuilder;
+use FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\QueryBuilder\QuoteQueryJoinQueryBuilderInterface;
 use FondOfOryx\Zed\ProductListSearchRestApi\ProductListSearchRestApiDependencyProvider;
 use Orm\Zed\ProductList\Persistence\SpyProductListQuery;
 use Spryker\Zed\Kernel\Persistence\AbstractPersistenceFactory;
@@ -11,7 +15,6 @@ use Spryker\Zed\Kernel\Persistence\AbstractPersistenceFactory;
 /**
  * @codeCoverageIgnore
  *
- * @method \FondOfOryx\Zed\ProductListSearchRestApi\Persistence\ProductListSearchRestApiRepositoryInterface getRepository()
  * @method \FondOfOryx\Zed\ProductListSearchRestApi\ProductListSearchRestApiConfig getConfig()
  */
 class ProductListSearchRestApiPersistenceFactory extends AbstractPersistenceFactory
@@ -30,5 +33,21 @@ class ProductListSearchRestApiPersistenceFactory extends AbstractPersistenceFact
     public function createProductListMapper(): ProductListMapperInterface
     {
         return new ProductListMapper();
+    }
+
+    /**
+     * @return \FondOfOryx\Zed\ProductListSearchRestApi\Persistence\Propel\QueryBuilder\ProductListSearchFilterFieldQueryBuilderInterface
+     */
+    public function createProductListSearchFilterFieldQueryBuilder(): ProductListSearchFilterFieldQueryBuilderInterface
+    {
+        return new ProductListSearchFilterFieldQueryBuilder($this->getConfig());
+    }
+
+    /**
+     * @return \FondOfOryx\Zed\CartSearchRestApi\Persistence\Propel\QueryBuilder\QuoteQueryJoinQueryBuilderInterface
+     */
+    public function createQuoteQueryJoinQueryBuilder(): QuoteQueryJoinQueryBuilderInterface
+    {
+        return new QuoteQueryJoinQueryBuilder();
     }
 }
