+++ bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/Persistence/ProductListSearchRestApiRepositoryInterface.php
@@ -11,7 +11,7 @@ interface ProductListSearchRestApiRepositoryInterface
      *
      * @return \Generated\Shared\Transfer\ProductListCollectionTransfer
      */
-    public function searchProductList(
+    public function findProductList(
         ProductListCollectionTransfer $productListCollectionTransfer
     ): ProductListCollectionTransfer;
 }
