--- bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/ProductListSearchRestApi/Persistence/ProductListSearchRestApiRepositoryInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\ProductListSearchRestApi\Persistence;
-
-use Generated\Shared\Transfer\ProductListCollectionTransfer;
-
-interface ProductListSearchRestApiRepositoryInterface
-{
-    /**
-     * @param \Generated\Shared\Transfer\ProductListCollectionTransfer $productListCollectionTransfer
-     *
-     * @return \Generated\Shared\Transfer\ProductListCollectionTransfer
-     */
-    public function findProductList(
-        ProductListCollectionTransfer $productListCollectionTransfer
-    ): ProductListCollectionTransfer;
-}
