--- bundles/product-list-search-rest-api/src/FondOfOryx/Zed/ProductListSearchRestApi/ProductListSearchRestApi/ProductListSearchRestApiDependencyProvider.php
@@ -1,85 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\ProductListSearchRestApi;
-
-use Orm\Zed\ProductList\Persistence\SpyProductListQuery;
-use Spryker\Zed\Kernel\AbstractBundleDependencyProvider;
-use Spryker\Zed\Kernel\Container;
-
-/**
- * @codeCoverageIgnore
- */
-class ProductListSearchRestApiDependencyProvider extends AbstractBundleDependencyProvider
-{
-    /**
-     * @var string
-     */
-    public const PROPEL_QUERY_PRODUCT_LIST = 'PROPEL_QUERY_PRODUCT_LIST';
-
-    /**
-     * @var string
-     */
-    public const PLUGINS_SEARCH_QUOTE_QUERY_EXPANDER = 'PLUGINS_SEARCH_QUOTE_QUERY_EXPANDER';
-
-    /**
-     * @param \Spryker\Zed\Kernel\Container $container
-     *
-     * @return \Spryker\Zed\Kernel\Container
-     */
-    public function provideBusinessLayerDependencies(Container $container): Container
-    {
-        $container = parent::provideBusinessLayerDependencies($container);
-
-        return $this->addSearchQuoteQueryExpanderPlugins($container);
-    }
-
-        /**
-         * @param \Spryker\Zed\Kernel\Container $container
-         *
-         * @return \Spryker\Zed\Kernel\Container
-         */
-    public function providePersistenceLayerDependencies(Container $container): Container
-    {
-        $container = parent::providePersistenceLayerDependencies($container);
-
-        return $this->addProductListQuery($container);
-    }
-
-    /**
-     * @param \Spryker\Zed\Kernel\Container $container
-     *
-     * @return \Spryker\Zed\Kernel\Container
-     */
-    protected function addSearchQuoteQueryExpanderPlugins(Container $container): Container
-    {
-        $self = $this;
-
-        $container[static::PLUGINS_SEARCH_QUOTE_QUERY_EXPANDER] = static function () use ($self) {
-            return $self->getSearchQuoteQueryExpanderPlugins();
-        };
-
-        return $container;
-    }
-
-    /**
-     * @return array<\FondOfOryx\Zed\ProductListSearchRestApiExtension\Dependency\Plugin\SearchQuoteQueryExpanderPluginInterface>
-     */
-    protected function getSearchQuoteQueryExpanderPlugins(): array
-    {
-        return [];
-    }
-
-    /**
-     * @param \Spryker\Zed\Kernel\Container $container
-     *
-     * @return \Spryker\Zed\Kernel\Container
-     */
-    protected function addProductListQuery(Container $container): Container
-    {
-        $container[static::PROPEL_QUERY_PRODUCT_LIST] = static function () {
-            return SpyProductListQuery::create();
-        };
-
-        return $container;
-    }
-}
